201 Repo/Appster/User.java

public class User implements Person {
	private String userName;
	private String userPassword;
	
	public User(String userName, String userPassword) {
		this.userName = userName;
		this.userPassword = userPassword;
	}

	@Override
	public boolean authenticate(String name, String password) {
		return userName.equals(name) && userPassword.equals(password);
	}

	public String getUserName() {
		return userName;
	}

	public void setUserName(String userName) {
		this.userName = userName;
	}

	public String getUserPassword() {
		return userPassword;
	}

	public void setUserPassword(String userPassword) {
		this.userPassword = userPassword;
	}
}